+++ tests/test_lin_rg.py
@@ -4,8 +4,8 @@ import pytest
 from numpy.testing import assert_allclose
 from numpy import ones, zeros, float64, array, append, genfromtxt
 
-from touvlo.lin_rg import (normal_eqn, cost_func, reg_cost_func, grad,
-                           reg_grad, predict, h)
+from touvlo.supv.lin_rg import (normal_eqn, cost_func, reg_cost_func, grad,
+                                reg_grad, predict, h)
 from touvlo.utils import numerical_grad
 
 TESTDATA1 = os.path.join(os.path.dirname(__file__), 'data1.csv')
