+++ touvlo/supv/nn.py
